requirements.txt
opencv-python==4.8.1.78
numpy>=1.26.4
onnxruntime>=1.17.0
scipy>=1.12.0
tqdm>=4.66.1
mediapipe==0.10.21
paho-mqtt>=1.6.1
